refactor(ff-a): implementation of FFA_MSG_WAIT v1.0
The implementation of FFA_MSG_WAIT/FFA_MSG_POLL v1.0 was still
present in code common to hypervisor and SPMC.

SPMC's implementation with recent work on indirect messaging
and runtime models, now matches FF-A v1.1 specification.

Thus, FF-A v1.0 FFA_MSG_WAIT/FFA_MSG_POLL is needed only to
avoid breaking the hypervisor's functionality, and some old
tests.

The referred FF-A v1.0 behavior was factored out of the SPMC
codebase, and is now specific to the hypervisor.

Change-Id: If7a9bd6fbfd19c0692371e430e29721c34ed89dd
Signed-off-by: J-Alves <joao.alves@arm.com>