refactor: implementation of HF_MAILBOX_*_GET interfaces
The interfaces HF_MAILBOX_WRITABLE_GET and HF_MAILBOX_WAITER_GET
are to accompany the FF-A v1.0 indirect message implementation.

This functionality is not used in the context of the SPMC,
hence factoring it out of common source files to SPMC
and hypervisor. This is now place in an hypervisor-specific
file.

Change-Id: I81332c286ad8a48fbe9a2425023d94ed7520fee8
Signed-off-by: J-Alves <joao.alves@arm.com>