ci: omit assertions disable build and run
This change [1] added a configuration in build.sh to build and run with
assertions disabled. This caused a significant overhead on the jenkins
job because now hafnium is built twice and the same qemu test suite run
8 times. This largelly increase the overall developer cycle test time
and also increases the possiblity to hit CI infra failures.
Revert this config for now with the intent to create a new dedicated job
config for disabled assertions. The TF-A CI already covers the disabled
assertions build.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/14949/11/kokoro/build.sh#59

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I302527e2a27115d7349b43622ddf19a5edb5a090