fix(misc): allow image origin address up to 48 bits
The ADRP AArch64 asm instruction permits offsets of up to 4GB.
The image origin address calculation done hypervisor and VM entry code
path does not permit relocating an image to high adddresses.
Use the appropriate relocation symbols to permit loading the Hafnium
image to an address of up to 48 bits.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I3e7e74dfbe7aa592cea30ecb1bb738cf2078ac2c