test(interrupts): secure interrupt target in RUNNING state
This patch adds a new test to exercise secure interrupt targeted
for an SP in RUNNING state. The test is made part of a new test suite
named secure_interrupts.

All the tests in this test suite are run as part of two different
configurations: one where the target SP is a S-EL1 partition and
another where the target SP is a S-EL0 partition.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: Ibdea25bffa3c8285509e1c21a2ff93f2cc0acdda