test(interrupts): secure interrupt target in PREEMPTED state
This patch adds a new test to exercise secure interrupt targeted
for an SP in PREEMPTED state. The test is made part of a new test suite
named secure_interrupts.

Also, the action in response to a non-secure interrupt is updated for
the second SP such that it is preempted.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: Icd87ff4b8e73f006801edaf2df246cd52198726d