feat(spmc): implement FFA_PARTITION_INFO_GET_REGS
Provide basic implementation of FFA_PARTITION_INFO_GET_REGS. This
implementation returns information of all partitions if UUID queried is
NULL or for any partitions that match the provided UUID, via registers,
as opposed to rx/tx buffers.
This implementation does not support the use of tags and does not
support forwarding the calls from the hypervisor to the spmc in setups
where hafnium is both hypervisor and SPMC. This is expected to be
implemented in future patches.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I66155de8924318f8df29bb67a8b268d71cdf4dcc