test(interrupts): exercise managed exit
Added a new test in `interrupts` test suite to exercise managed exit
for an S-EL1 partition.

Made the necessary update to the fields in manifest of the first
service partition to request SPMC to signal managed exit through
vIRQ pin.

Since this new test shares a lot of functionality with an existing
test, the common code has been refactored into helper functions.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: Ic6051939d484d716ea1e81c519f4008c1e7f0926