feat: turn first SP vcpu off upon physical PE off
For a S-EL1 SP registering the cpu off power management event (upon
receiving the event from the SPMD on PSCI CPU OFF), turn the
corresponding pinned vCPU to off.
The implementation currently limits handling of power management
messages to S-EL1 MP SPs. In a later iteration, the SP vCPU will have to
be resumed and conveyed the power management message prior to the above
action.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I3da387c30cabe737a72da600687c9161c7e2c938