feat(htest): provide test skip helper
Re-use the recently introduced test preconditions to define a macro
permitting to skip a test from a test suite.
This permits keeping the code built but actually not run the test
in the CI.
It remains possible to run the tests marked to skip by using the
"--test" hftest option.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I5228800da4927c9ce2d9ad924632456f78dbba6d