ci: update bl31 (no spmd) prebuilt to TF-A v2.8
As Trusty tests were removed from test runs [1], there is no longer a
dependency to BL31 image built with Trusty SPD and the downstream
Android TF-A fork.
Refresh prebuilts with upstream TF-A. Those binaries are used with tests
run through the test.sh script (either qemu or FVP platform) where the
SPMD is not required.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/16826

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I0c5c3b94eb4829bdf90d186039eeb754fc9a48ce