ci: QEMU only uses option 'cpu' set to 'max'
The QEMU supports the emulation of the cortex-a53, which was originally
adopted in the execution of the Hafnium tests.
The Hafnium project has been repurposed primarily for the SPMC reference
implementation, in which VHE is a critical feature. The cortex-a53
doesn't support VHE. As such, this patch removes it from QEMU test runs.

Change-Id: Ie066bccf25e6173cadaf2f2e5d131aa6080b7e02
Signed-off-by: J-Alves <joao.alves@arm.com>