feat(hftest): add third SP for ffa_secure_partitions setup
This patch adds a third secure partition in the `ffa_secure_partitions`
setup. This is needed to exercise a new functionality added in the
Hafnium to queue secure interrupts targetted for other secure
partition. The test itself is added in further patch.

Also, corresponding changes have been made to the `ffa` test suite to
accommodate third secure partition.

Change-Id: I40ecb61415bd734aa1c87ce973b9616671cd9add
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>