feat(hftest): add arguments to hftest.py to aid debug
Add arguments to 'hftest.py' to aid with debugging or
system development:
- "--show-output": hftest default is to capture model's stdout
and dump it in a file, without any visual feedback as execution
is happening. This option makes it such we can see the logs of
the on-going tests.
- "--debug": with this argument hftest adds the arguments to
the model call, to allow developer to connect debugger (gdb or
Arm DS).

Currently functionality is working for FVP platform. QEMU is WIP.

Change-Id: I62b5f82b50829dc541ee7a47753996409265e44b
Signed-off-by: J-Alves <joao.alves@arm.com>