fix: update TPIDR_EL2 on switch to normal world
Following [1] fix an issue in that TPIDR_EL2 is not updated when a PE
exits to the normal world. This is to cover the particular the case of
when the SPMC is entered from the normal world, processes an FF-A
interface and returns straight to the normal world without resuming a
vCPU in the secure world. Fix by updating TPIDR_EL2 with the address of
the 'other world VM' vCPU matching the PE on which the SPMC is exiting.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/16767

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: Ifa5a90ee69b6c0d8d2fdc41fcbf033e3834990e8