refactor: remove absent prng
The absent implementation of prng was removed. Now the only fake prng
implementation will be used.

The current implementation of prng is very simple and must be
re-implemented in a production system.

The absent implementation for prng was used before for some platforms.
In this case, the branch protection and MTE features (if enabled) were
initialized with zero value. This change is needed to have a non-zero
value from the prng function in any case.

Change-Id: I19bf2b100a3adc18e3d5d262909c7568c1c5f550
Signed-off-by: Dmitrii Martynov <fkr655@yandex.ru>