refactor(interrupts): precedence for action in response to interrupts
While winding the call chain in response to a direct message, SPMC
must follow the precedence for action in response to non-secure and
other secure interrupt before resuming the target vCPU.

This patch simplifies the logic used to implement the precedence
rule by conditionally increasing the interrupt threshold (i.e.
numerically decreasing the priority mask) by checking the action
of the target VM.
This also simplifies the logic while unwinding the call chain.

Change-Id: I7537ce13fe5e80cd2babb4d082cfc171dc9a24f0
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>