ci: add tests with both worlds for EL3 SPMC
Added new test setup to validate the implementation of the EL3 SPMC. The
tests are based on 'ffa_secure_partitions', and use the same message
loop logic. The new setup is added to kokoro/test_el3_spmc.sh, such that
the tests are executed as part of the CI job dedicated to EL3 SPMC.

The tests include:
* Direct messaging functionality.
* Partition info get interface.

The same partitions are used to run with Hafnium as the SPMC, to provide
a means to infer about feature parity. This configuration can be
executed by calling kokoro/test_spmc.sh and providing --parity argument.

Signed-off-by: Shruti Gupta <shruti.gupta@arm.com>
Signed-off-by: Karl Meakin <karl.meakin@arm.com>
Change-Id: Idc6a6132fc61693bb7e1e6fcc9a78089c5408f7a