feat: add security state metadata to S2 PTEs
Use an additional SW bit in S2 PTEs to specify whether an FF-A memory
region is secure or non-secure.
With TZ/S-EL2, this information consumed by the HW is held in S1 PTEs.
This lets the SPMC track the security state of a mem region through a
SP's S2 PTs without need to peek in S1 PTs.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I2faca658ccdfb2d01197957177d3d09e016c4bcb