ci: update load lib path to accomodate LLVM
For running unit tests, the host dynamic linker load lib path variable
must be adjusted to provide the LLVM toolchain lib path (in particular
to the C++ shared library). The toolchain lib path is changing somewhere
from clang 12 to clang 15 hence provide a more dynamic way to find out
this correct path so that it works for older and newer toolchains.

Change-Id: Ib904db819901633e1cb4c6dcacc9ad9435c19e99
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>