refactor(hftest): simplify stdout buffer logic
Reduce stdout buffer size. When buffer is full, always restart at
the beginning of the buffer. This reduces the chance of corner cases
and issues related to calculating the count of characters to write.

Also add a mechanism to flush the stdout buffer at the end of a dlog
call from (S)EL1. From (S)EL2, this call will be noop as the stdout
buffer is specific to VMs/SPs only.

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: I0d643731dfe0916013c92bae8525671700db11be