feat(spmc): support for el3 direct messaging
This patch makes changes required to support el3 direct messaging from
an el3 spmd logical partition. It adds the logical partition
identifiers that are a reserved range as defined in the patch. The patch
also changes the checks for valid direct requests and responses to be
accepted or targeted to the spmd logical partition. When a direct
request is received from an spmd logical partition, the target secure
partition is put in spmc scheduled mode with all interrupts masked. When
a direct response is targeted to an spmd logical partition, the spmc
switches back to el3.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I11a6ac1021e3e8913afc0f64cbbe89742b73fe73