fix(memory share): clear memory operation
Through the memory sharing process there is the possibility either
sender or retriever requests the SPMC to clear the memory.
To do so, the SPMC maps the region into its own address space, and
clears the memory.
In case the region is non-secure there was no way to detect whether
the region was non-secure other than infering it from the owner's
ID.
Now the security state is tracked in all PTEs, recurring to the NS
bit.

Hafnium was already saving the owner's original mode during the
handling of memory share/lend/donate, after retrieving it from
the PTEs. The mode contains the information about the region's
security state.

This patch removes all inferences to memory's security state
based on owner's ID, and uses the owner's original mode.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ia38068adff3f237f5aa72481fc301a5555e6be1f