test: disallow migrating a blocked SP to another CPU
This patch adds a test to make sure the vCPU of an UP SP cannot be
migrated from current CPU to a different physical CPU while the vCPU
is in BLOCKED state as part of an SP call chain.

This patch also adds a helper function to check if the receiver
endpoint is an UP SP. This helper is used as a precondition for the
newly introduced test.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I018c152737dbd9abae7385e8858a8d231d5e82e8