fix(hftest): fix parsing the test json in hftests
Printing to the UART prior to hftests parsing the JSON structure
of the available tests caused the parsing to fail and the tests
to stop. Wrap the JSON structure in hftest_ctrl tags which the
hftest script can use to identify which lines of the uart contain
the JSON structure.

This is important as future patches will introduce the ability to
run functions during an SPs initialsation stage and any prints in
this code would cause the HFTEST to crash without the hftest_ctrl flags.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Ib63a06f876e80cd4e4e360282bffda2c68474346