feat: implement stricter checks for state transition
Currently, asserts are in place to check the expected tentative state
of the current execution context upon a legal state transition under
the appropriate partition runtime model. The `next_state` variable
which tracks the tentative state is prepopulated.

This allows for a hypothetical bug to go unnoticed. For example, if
upon an illegal transition, the next_state is left unmodified, the
asserts would not fail due to the prepopulated value.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I102ec4470eae4b1fea90063fd334e2769f00cb9e