feat(uuids): add support for multiple UUIDs per partition
Add manifest parsing, population of load-time data structures,
and manifest unit tests needed for supporting usage of multiple
UUIDs per partition.

This patch does not add support to ffa_partition_info_get
for discovery of multiple UUIDs/partition. That will be
addressed in a subsequent patch and for now, only the first
UUID for a partition is reported in the partition info
descriptors.

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: I23f6c21dfab2c24cb4a3f7e664616b7bd183b88d