chore(logs): clean up logs for aborted VMs
When an SP was attempted to be run via ffa_run or direct
messsaging interfaces a check is performed by reading the vm->aborting
flag. Previously, a log stating "Aborting VM" was printed along with
this check. This can become redundant if multiple requests to an
aborted VM are made. Instead, print "Aborting" message only when
setting the VM abort flag and print a more specific error message when
check is performed and verbose logs are set.

Also add flag to sp_boot_next parameters to indicate whether previous
SP was initialized successfully or not and prevent printing
"Initialized VM" when SP was actually aborted.

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: I84c50c2b11aef12544cfcc66999b1aedfdaad781