refactor: partition type based on FF-A ID
Make the helpers for the detecting the type of
FF-A partition part of the vmapi/ffa.h header.
This is because it is a concept defined in the
specification, and it is common to all FF-A implementations.

This patch drops the plat_ffa_is_vm_id implementation
and creates a new 'ffa_is_vm_id'.
Also replaces all instances of the former
with an invocation to the later.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: If46f83f7943ed669a333c52d6763f90553f7d5b5