test: add set up stage for to hftest services
Add the HFTEST_SERVICE_SET_UP macro which will run the function
during the initialisation stage of a service partition. This is the
time when the SP's primary vCPU is in execution context initialisation
just before the SP uses FFA_MSG_WAIT to terminate the boot sequence,
as specified in FF-A's boot protocol.

A service must specify the set up functions it wants to run in the
hftest-service-setup node in it's partition manifest.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I33464dd3454137eb3a1bf7678cafc75f16c8c5de