fix: correct el for primary_only_ffa test config
Primary VMs must be hosted in the normal world. For the primary_only_ffa
configuration this was not the case, the primary manifest for the
VM set the exception-level to S-EL1 which caused a check in
load_primary to fail. Correct this and fix the entrypoint offset so
the vm boots and runs successfully.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I0556187b7fe2f0bfa1d7abf9895d5e0f8dcad8b3