feat(ff-a): add multiple uuids to partition info desc
FF-A v1.2 spec allows for a partition to list multiple UUIDs.
This requires changes to how FFA_PARTITION_INFO_GET fills the
partition information descriptors it returns.

As per FF-A v1.2 section 6.2.3 Partition UUID usage,
FFA_PARTITION_INFO_GET should return an entry corresponding to each UUID
in the returned partition information.

Additionally, if the Nil UUID is specified, each valid VM-UUID pair
contiributes to the total returned count. In other words, the returned
count is equal to the sum of the products of the VMs and the number of
UUIDs that they export.

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: Ibd7e1cbe9f819a24c4eb4988f05f8360cf623e7b