fix: ensure SVE traps are enabled when running SPs
Add CPTR_EL2 to the VM's hypervisor state. CPTR_EL2 is restored when a
vCPU is scheduled such that traps are properly set when a vCPU runs.
When a vCPU exits to SPMC, CPTR_EL2 is restored to an Hypervisor/SPMC
defined value. This is mainly to offer different trap states while
running PVM vCPUs versus secondary VMs/SPs vCPUs e.g. this change
permits SVE to be used by the PVM but not secondary VMs/SPs.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I1afeb1d0e166103692af830a3e4ba9d80157a4f8