feat(memory share): protect on memory lend/donate
For memory lend/donate from NWd into the SWd, in which all borrowers
are SPs, the SPMC uses the `enum ffa_map_action` to request
the monitor to configure the memory as secure, upon updating sender's
PTs.

If the memory is protected successfully, this is tracked in the
`share_state` structures, such that the appropriate unprotect call
can be taken upon memory reclaim. It guarantees that the PAS is
reestablished to that configured by the SPMC upon lend/donate call.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I8a9c674e8bfecf5db8768a4ec08c6925bded110c