feat(hftest): macro to declare a long running test with precondition
This patch provides a macro to declare a long running test with a
precondition function.

One of the existing test, disallow_migration_blocked_sp, needs longer
time to run on some aarch64 developer machines as it has explicit
wait time to allow for call chain to be established on boot CPU.

Hence, this test now leverages the newly introduced test macro. Also,
the timeout for long running tests has been increased to 100s.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I93e536b95601adb4a342207046d2242fe7cd7395