test(memory share): add test for FF-A v1.1 endpoints
Add tests for sharing memory to/from a v1.1 endpoint.
v1.1 is worth testing separately as it is the first version
where the `memory_access_desc_size` field was added to the
`ffa_memory_region` struct. The v1.1 `ffa_memory_access` struct
does not contain the impdef field added in v1.2 so it tests
that the spmc can handle passing around a smaller
memory access descriptor.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Ibfd6d4d12e1ffb862b44a48c1ca1ace62a0e3d20