refactor(memory share): split check of hyp retrieve request
Section 17.4.3 of the FF-A v1.2 ALP0 spec states that the SPM should
distinguish between retrieve request originating from the hypervisor and
one originating from a Borrower VM, forwarded by the Hypervisor by
checking the endpoint memory access descriptor count. For the former
case the count should be 0 otherwise >=1. This patch separates this
check from the validation that the all other fields in the transaction
descriptor, except the handle, are 0.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I1829069450633db560048cf0bc08d1ad055a8dcc