fix(simd): mask out SMCCC SVE hint bit
The SMCCC SVE hint bit is relevant to the SPMC when an SMC is propagated
from a caller in the normal world. For an SMC/HVC originating from a VM
or SP vCPU, mask out the SVE hint bit as Hafnium only supports context
switching Adv. SIMD for VM/SP vCPUs. There's no reason for a VM or SP
to set the hint bit as VMs/SPs do not support SVE/SME.

Change-Id: Ic1462816cecd953dba7589d511c43aca74c35cac
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>