test(ff-a v1.2): ffa_run to zero extended registers
FF-A v1.2 introduces an extended set of available registers
for its interfaces. When a partition that is handling a
direct request yields, it may be resumed by ffa_run. In order
to maintain the full extended register state when passing execution
back to the yielding partition, allow ffa_run to zero the extended
register set to preseve runtime state of the calling partition.

Patch also includes tests demonstrating this use case.

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: I27a01ab40f9f74867536812c24a0d665e782754f