test: exercise interrupt handling with interrupts masked
This patch adds a test to exercise secure interrupt handling while
the Secure Partition runs with interrupts masked. This test helps to
validate the functionality of the SPMC to intercept a direct response
message if there are pending virtual secure interrupts and reschedule
the partition to handle the pending interrupt.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I6ae5be1a4a3594c40c4582d19f25be0483183541