fix(interrupts): reset implicit completion signal
The implicit_completion_signal field is necessary for performing
necessary bookkeeping, for example, when the SP gets preempted by a
Self S-interrupt.

However, it is not needed in secure interrupt runtime model because
an SP cannot be preempted by any interrupt in this runtime model.

This patch resets this field when SPMC runs the execution context of
the target SP for handling the pending virtual interrupt.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I3075aeed7c822fe2019647747c6a01b6bc7abecc