refactor(memory share): add helper for accessing receivers
FF-A v1.1 introduces the memory access descriptor size field to
memory transaction descriptors with the intention that these will help
maintain forwards compatibility of the `ffa_memory_access` structs in
future versions. To achieve this we introduce a helper function for
accessing a receiver from the receiver array which uses
`memory_access_desc_size` to fetch this.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Ic36f83093c9061100dfa293f0140469b0398aadc