feat(memory share): protect/unprotect memory
Use the protect/unprotect functions in ffa_region_group_identity_map.
The referred functions are responsible for updating S2 PTs.
The functions have been extended to use the map actions to extended
the previous uses for "check" and "commit" the page tables.

The "check" is extended with "check protect" to be able to request
the monitor to configure memory into secure PAS.
The "commit" is extend with "commit unprotect" to be able to reset
memory back to non-secure PAS.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ifd209cea2d1d351698e87d9e8796daa376c28b90