test(dir_msg): add FF-A version based tests
Test the usage of FFA_MSG_SEND_DIRECT_REQ2 based on the sender
and receiver endpoints' respective FF-A version.

To enable these tests, add a Secure Partition using FF-A
v1.1 for backwards compatibility testing.

Also change service3 messaging method permissions to not
support FFA_MSG_SEND_DIRECT_REQ2 to test that it cannot
receive such a request if it is v1.2 but the messaging
method is not declared in its manifest.

Test cases:
 - V1.2 endpoint sending a direct request via FFA_MSG_SEND_DIRECT_REQ2
   to another v1.2 endpoint that doesn't support this ABI shall fail
 - V1.2 endpoint sending an FFA_MSG_SEND_DIRECT_REQ2 to a v1.1 endpoint
   shall fail
 - V1.1 using FFA_MSG_SEND_DIRECT_REQ2 shall fail

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: Icad7608a083340d3f55a13054d6b02981d26a452