fix(memory share): fix behaviour of relinquish clear flag
The `Zero memory after relinquish` flag set in a relinquish
request should take priority over the value of the
`Zero memory flag` set in the retrieve request previously.
This is not currently the case and this patch corrects that.

We also correct the tests to reflect this as well as adding some
additional tests to ensure in cases where the flag's value during
the retrieve request does not match the flag during the relinquish
request, the behaviour that occurs is dictated by the relinquish
request.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I770362374da9c333952249f30e17bf34edd83d6c