test(memory share): device lend SP to SP
Adds two tests for lending device memory between SPs. The first
checks that when an SP lends device memory to another SP it
looses access to it during this period. The second checks that
the lent device memory can be accessed by the borrower and
then once reclaimed the original owner regains access and the
borrower no longer has it. The second test uses the UART1 memory range
so the output caused by the test can be checked in the uart1 log.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: If7a49b338b526c79482a14ce999ea248121c9e7f