refactor: simplify interception of FF-A calls
Reduced the duplication of code in the interception
of FF-A calls.

The interception of FFA_MSG_WAIT and FFA_MSG_SEND_DIRECT_RESP
in case there is secure interrupt pending is at the end
of handling both ABIs.

As a result, both would have determined what the next vCPU
to execute, without considering the secure interrupt.

The function `plat_ffa_intercept_call` sets the next vCPU
in a preempted state, and returns execution back to the
caller with eret FFA_INTERRUPT.

The implementation of plat_ffa_msg_wait_prepare was also too
convoluted. Refactored to handle based on runtime model.

BREAKING: The preempted_by_secure_interrupts test now fails
for S-EL0 partitions. Fix in following tests.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I20e3e2d82ce15e542243a457e31e1a13ff5c6128