refactor: set vcpu of the NWd in preempted state
When the SP invokes FFA_MSG_WAIT, the function
`plat_ffa_preempted_vcpu_resume` can manage the vCPU state
both for when NWd is preempted or another SP.
The FFA_NORMAL_WORLD_RESUME ABI is populated onto the context
of the vCPU.
So they will be propagated onto the NWd at context restore
after handling FFA_MSG_WAIT from the SP that was targeted by
the interrupt.
Set the NWd vCPUs in preempted state similarly to what we
do for SPs.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I202c0076a79889d73aac3264c69c334b680fbbee