feat(memory share): validate sp device mem ranges
Validate the device memory specified in a partitions manifest
belongs to the ranges given in the SPMC manifest. This patch
also introduces the ability to define ns device memory ranges
in the SPMC which are needed for later enablement of NWd to SWd
device memory sharing.

Tests made to fail in manifest_test.cc are fixed in this patch
and an additional test is added to check the validation works
as expected. In addition the test checking that normal memory
regions and device memory regions cannot overlap has been removed
as these memory regions need to be allocated in the regions defined
in the SPMC so this would fail with MANIFEST_ERROR_MEM_REGION_INVALID
and is checked by the test added in this patch.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: If93418614f9b58e61e625839ed689048e3c75a03
Signed-off-by: J-Alves <joao.alves@arm.com>