fix(memory share): broken assert in send from SP to multiple
An assert in ffa_send_check_transition failed when sharing
v1.1 memory access descriptors as it made assumptions about
the size of the memory access descriptor struct. This patch
changes to using the ffa_memory_region_get_receiver helper which
uses the memory_access_descriptor_size field to prevent this.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: If40ca7192ace85dedd602f22c5dea5d7ec87d100