fix(mem sharing): return error if receiver count is 0
Previously we were only checking the receivers count in the memory
region descriptor was less than the maximum share recipients but not
that it wasn't 0. This meant the assert on line 581 of ffa_memory.c
could fail if the receiver count was 0. To avoid tripping the assert,
and instead gracefully return FFA_ERROR in this case, expand the
previous receiver count check to include the 0 case.

This patch also tidies up the accessing the receivers for v1.0 memory
access descriptors.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I81a53247e52333e17451773cf22f89f06359b9fd