fix(ff-a): report SRI unsupported to secure partitions
The `FFA_FEATURES` ABI allows querying support for the "schedule
receiver" interrupt on FF-A version 1.1 or greater from the normal
world. This is useful for the hypervisor or OS kernel running as the
normal world, but not for secure partitions. Therefore, return
`FFA_NOT_SUPPORTED` if the query is from an SP.

Change-Id: Id7989d749963b7343432452b5071f8369bc90338
Signed-off-by: Karl Meakin <karl.meakin@arm.com>