fix: partition message offset check
FF-A v1.2 ALP1 Table 7.2 defines a message offset field for specifying
the message payload location within the RX/TX buffer.
In Hafnium implementation and tests, there is an assumption this offset
is fixed to the size of the partition message header. It means the
message payload always follows the header. Make it explicit in the ABI
by rejecting an FFA_MSG_SEND2 call when the message offset field is
different from the size of the partition message header.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I26b93eef9fa506ffd8cbf5dd51075d8c882b58ce