fix: fix typo in static_assert
The assertion `static_assert(sizeof(((struct
ffa_memory_region){0}).receiver_count == 4))` was intended to compare
the size of the `receiver_count` field against 4. But due to a typo, it
actually compares the *value* of the `receiver_count` field against 0.
`sizeof` then returns the size of the type of that expression (ie
`sizeof(typeof(0 == 4))` => `sizeof(int)`), which is guaranteed to be
nonzero. In other words, this static_assert will always be true,
regardless of the size of the `receiver_count` field, so a breaking
change would silently pass undetected.

This patch fixes the assertion to correctly compare the size of the
`receiver_count` field against 4.

Change-Id: Ieb08629b4ac5c14cffed1d0d88853039db552b91
Signed-off-by: Karl Meakin <karl.meakin@arm.com>