fix(direct message): reorder checks for FFA_BUSY
In case the FF-A endpoints is in another direct message request
the specification mandates the error code to return shall be
FFA_BUSY.
This was conflicting with the runtime model checks, that returned
FFA_DENIED if not compliant.
Runtime model checks were moved until after the checking whether SP
is busy.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I6d50863e9004525d469c62c109fab8c7ad9bbb2f