feat(ff-a): use memcpy that is trapped GPF
In `FFA_PARTITION_INFO_GET` the SPMC copies the paritition info
descriptor to the RX buffer of the NWd. This is prone to a GPF,
if NWd request realm to add memory to the realm PAS.

With `memcpy_trapped` the SPMC should be able to recover from GPF,
and return error FFA_ABORTED to the caller.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I2b25276c36eb70a46290f7157d71b3df9f1a4adb