feat(dlog): check arguments to `dlog` at compile-time
Annotate the `dlog` function with the `format(printf)` attribute. This
means arguments to `dlog` will be checked for correctness at
compile-time by clang or GCC (see
https://clang.llvm.org/docs/AttributeReference.html#format and
https://gcc.gnu.org/onlinedocs/gcc/Common-Function-Attributes.html).

Had to disable the `PREFER_PRINTF` check in `checkpatch` because clang
does not recognize `__printf(1, 2)` as a replacement for
`__attribute__((printf(1, 2)))`.

BREAKING CHANGE: Incorrect uses of `dlog` will now be a compile-error

Signed-off-by: Karl Meakin <karl.meakin@arm.com>
Change-Id: Ib1f1968d4d90078ab3031b5e7733ec7cf7c95575