fix(ff-a): check that hypervisor's TX buffer is set-up
The `api_get_rxtx_description` function reads from the hypervisor's TX
buffer in the case of a forwarded RXTX_MAP call, but it is possible to
make a forwarded RXTX_MAP call without first setting up the hypervisor's
RX/TX buffers. This would result in a null-pointer dereference.

This patch fixes this issue by returning FFA_INVALID_PARAMETERS if the
hypervisor's TX buffer is not set-up.

Change-Id: Iecd7fb39a266270465fc627908acd5b811c80ddc
Signed-off-by: Karl Meakin <karl.meakin@arm.com>